ginkgo: audio: Add fast input profile for the record concurrency
If two recording apps both select the primary input with
fast flag, second app will fail to start because the hal
doesn't support two inputs with the same use case if input
can't be reused since of the limit of primary input's max
open count.

Add the dedicated profile for “low-latency-record” to make
the first input can be reused if possible.

CRs-Fixed: 2550228
Change-Id: I233bd3eaf614b195df0a41402747f496d3a56b44